Move autoSkip related functions to separate file (#8691)
